Refactor table
